+++ smoke-test/tests/cypress/cypress/plugins/index.js
@@ -21,7 +21,7 @@ module.exports = (on, config) => {
   // `config` is the resolved Cypress config
 
   // eslint-disable-next-line global-require
-  require("./memoryUsage")(on);
+  require("./memoryUsageLogger")(on);
 
   // eslint-disable-next-line global-require
   require("cypress-timestamps/plugin")(on);
