+++ smoke-test/tests/cypress/cypress/plugins/memoryUsageLogger.js
